+++ mavlink-kotlin/src/main/kotlin/xyz/urbanmatrix/mavlink/connection/MavConnection.kt
@@ -15,25 +15,25 @@ public interface MavConnection : Closeable {
 
     @Throws(IOException::class)
     public fun <T : MavMessage<T>> sendV1(
-        systemId: Int,
-        componentId: Int,
+        systemId: UByte,
+        componentId: UByte,
         payload: T
     )
 
     @Throws(IOException::class)
     public fun <T : MavMessage<T>> sendUnsignedV2(
-        systemId: Int,
-        componentId: Int,
+        systemId: UByte,
+        componentId: UByte,
         payload: T
     )
 
     @Throws(IOException::class)
     public fun <T : MavMessage<T>> sendSignedV2(
-        systemId: Int,
-        componentId: Int,
+        systemId: UByte,
+        componentId: UByte,
         payload: T,
-        linkId: Int,
-        timestamp: Long,
+        linkId: UByte,
+        timestamp: UInt,
         secretKey: ByteArray
     )
 }
