+++ mavlink-kotlin/src/main/kotlin/xyz/urbanmatrix/mavlink/connection/stream/MavRawFrameReader.kt
@@ -26,12 +26,12 @@ internal class MavRawFrameReader(
             val payloadLength = inputStream.read()
             if (payloadLength == -1) throw EOFException()
 
-            when (versionMarker) {
+            when (versionMarker.toUByte()) {
                 MavFrameType.V1.magic -> {
                     val success = inputStream.advance(
                         MavRawFrame.SIZE_SEQ + MavRawFrame.SIZE_SYS_ID +
                                 MavRawFrame.SIZE_COMP_ID + MavRawFrame.SIZE_MSG_ID_V1 +
-                            payloadLength + MavRawFrame.SIZE_CHECKSUM
+                                payloadLength + MavRawFrame.SIZE_CHECKSUM
                     )
                     if (!success) {
                         drop()
@@ -45,7 +45,7 @@ internal class MavRawFrameReader(
                     val incompatibleFlags = inputStream.read()
                     if (incompatibleFlags == -1) throw EOFException()
 
-                    val signatureSize = if (incompatibleFlags == MavRawFrame.INCOMPAT_FLAG_SIGNED) {
+                    val signatureSize = if (incompatibleFlags.toUByte() == MavRawFrame.INCOMPAT_FLAG_SIGNED) {
                         MavRawFrame.SIZE_SIGNATURE
                     } else {
                         0
