+++ mavlink-kotlin/src/main/kotlin/xyz/urbanmatrix/mavlink/connection/tcp/TcpMavConnection.kt
@@ -33,8 +33,8 @@ public abstract class TcpMavConnection : MavConnection {
 
     @Throws(IOException::class)
     final override fun <T : MavMessage<T>> sendV1(
-        systemId: Int,
-        componentId: Int,
+        systemId: UByte,
+        componentId: UByte,
         payload: T
     ) {
         when (val s = state) {
@@ -50,8 +50,8 @@ public abstract class TcpMavConnection : MavConnection {
 
     @Throws(IOException::class)
     final override fun <T : MavMessage<T>> sendUnsignedV2(
-        systemId: Int,
-        componentId: Int,
+        systemId: UByte,
+        componentId: UByte,
         payload: T
     ) {
         when (val s = state) {
@@ -67,11 +67,11 @@ public abstract class TcpMavConnection : MavConnection {
 
     @Throws(IOException::class)
     final override fun <T : MavMessage<T>> sendSignedV2(
-        systemId: Int,
-        componentId: Int,
+        systemId: UByte,
+        componentId: UByte,
         payload: T,
-        linkId: Int,
-        timestamp: Long,
+        linkId: UByte,
+        timestamp: UInt,
         secretKey: ByteArray
     ) {
         when (val s = state) {
